A slight error handling improvement when the reference sequences don't match.
bug 21713

[git-p4: depot-paths = "//depot/software/bioinformatics/tools/pbh5tools/": change = 123908]

+++ src/python/pbtools/pbh5tools/CmpH5Merge.py
@@ -267,7 +267,8 @@ def cmpH5Merge(inFiles, outFile):
     except Exception, e:
         try:
             # remove the file as it won't be correct
-            os.remove(outFile)
+            if os.path.exists(outFile):
+                os.remove(outFile)
         except:
             pass
         raise
